// src/lib/api/types.d.ts
export interface operations {
// ...
  articleBySlug: {
    parameters: {
      query?: never
      header?: never
      path: {
        slug: string
      }
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: Record<string, unknown>
        content: {
          "application/json": {
            id: string
            /** @enum {string} */
            language: "id" | "en"
            title: string
            slug: string
            content: string
            excerpt: string
            metaTitle: string | null
            metaDescription: string | null
            /** @enum {string} */
            status: "published" | "draft" | "rejected" | "in_review"
            /** @enum {string} */
            visibility: "public" | "member"
            articleTranslationId: string
            featuredImage: string
            createdAt: string | null
            updatedAt: string | null
            topics: {
              id: string
              title: string
              slug: string
            }[]
            authors: {
              id: string
              name: string | null
              username: string
            }[]
            editors: {
              id: string
              name: string | null
            }[]
          } | null
        }
      }
    }
  }
// ...
  articleByLanguageInfinite: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody: {
      content: {
        "application/json": {
          /** @default 50 */
          limit?: number
          cursor?: string | null
          /** @enum {string} */
          language: "id" | "en"
        }
      }
    }
    responses: {
      /** @description OK */
      200: {
        headers: Record<string, unknown>
        content: {
          "application/json": {
            articles: {
              id: string
              /** @enum {string} */
              language: "id" | "en"
              title: string
              slug: string
              content: string
              excerpt: string
              metaTitle: string | null
              metaDescription: string | null
              /** @enum {string} */
              status: "published" | "draft" | "rejected" | "in_review"
              /** @enum {string} */
              visibility: "public" | "member"
              articleTranslationId: string
              featuredImage: string
              createdAt: string | null
              updatedAt: string | null
            }[]
            nextCursor: string | null
          }
        }
      }
    }
  }
  articleByTopicId: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody: {
      content: {
        "application/json": {
          /** @default 1 */
          page?: number
          /** @default 10 */
          perPage?: number
          /** @enum {string} */
          language: "id" | "en"
          topicId: string
        }
      }
    }
    responses: {
      /** @description OK */
      200: {
        headers: Record<string, unknown>
        content: {
          "application/json": {
            article: {
              id: string
              /** @enum {string} */
              language: "id" | "en"
              title: string
              slug: string
              content: string
              excerpt: string
              metaTitle: string | null
              metaDescription: string | null
              /** @enum {string} */
              status: "published" | "draft" | "rejected" | "in_review"
              /** @enum {string} */
              visibility: "public" | "member"
              articleTranslationId: string
              featuredImage: string
              createdAt: string | null
              updatedAt: string | null
            }
          }[]
        }
      }
    }
  }
  articleByTopicIdInfinite: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody: {
      content: {
        "application/json": {
          /** @default 50 */
          limit?: number
          cursor?: string | null
          /** @enum {string} */
          language: "id" | "en"
          topicId: string
        }
      }
    }
    responses: {
      /** @description OK */
      200: {
        headers: Record<string, unknown>
        content: {
          "application/json": {
            articles: {
              id: string
              /** @enum {string} */
              language: "id" | "en"
              title: string
              slug: string
              content: string
              excerpt: string
              metaTitle: string | null
              metaDescription: string | null
              /** @enum {string} */
              status: "published" | "draft" | "rejected" | "in_review"
              /** @enum {string} */
              visibility: "public" | "member"
              articleTranslationId: string
              featuredImage: string
              createdAt: string | null
              updatedAt: string | null
            }[]
            nextCursor: string | null
          }
        }
      }
    }
  }
// ...
  articleRelatedInfinite: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody: {
      content: {
        "application/json": {
          /** @default 50 */
          limit?: number
          cursor?: string | null
          currentArticleId: string
          /** @enum {string} */
          language: "id" | "en"
          topicId: string
        }
      }
    }
    responses: {
      /** @description OK */
      200: {
        headers: Record<string, unknown>
        content: {
          "application/json": {
            articles: {
              id: string
              /** @enum {string} */
              language: "id" | "en"
              title: string
              slug: string
              content: string
              excerpt: string
              metaTitle: string | null
              metaDescription: string | null
              /** @enum {string} */
              status: "published" | "draft" | "rejected" | "in_review"
              /** @enum {string} */
              visibility: "public" | "member"
              articleTranslationId: string
              featuredImage: string
              createdAt: string | null
              updatedAt: string | null
            }[]
            nextCursor: string | null
          }
        }
      }
    }
  }
// ...
  articleSearch: {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    requestBody: {
      content: {
        "application/json": {
          searchQuery: string
          /** @default 10 */
          limit?: number
          /** @enum {string} */
          language: "id" | "en"
        }
      }
    }
    responses: {
      /** @description OK */
      200: {
        headers: Record<string, unknown>
        content: {
          "application/json": {
            id: string
            /** @enum {string} */
            language: "id" | "en"
            title: string
            slug: string
            content: string
            excerpt: string
            metaTitle: string | null
            metaDescription: string | null
            /** @enum {string} */
            status: "published" | "draft" | "rejected" | "in_review"
            /** @enum {string} */
            visibility: "public" | "member"
            articleTranslationId: string
            featuredImage: string
            createdAt: string | null
            updatedAt: string | null
          }[]
        }
      }
    }
  }
// ...
}